+++ R/cancensus.R
@@ -288,7 +288,7 @@ VALID_LEVELS <- c("Regions","C","PR", "CMA", "CD", "CSD", "ADA","CT", "DA", 'EA'
 
 #' Query the CensusMapper API for available datasets.
 #'
-#' @param use_cache If set to TRUE (the dfault), data will be read from a temporary local cache for the
+#' @param use_cache If set to TRUE (the default), data will be read from a temporary local cache for the
 #'   duration of the R session, if
 #'   available. If set to FALSE, query the API for the data, and
 #'   refresh the temporary cache with the result.